Moved simular helpers in tests to utils
